Move MessageListItemWrapper to feature-specific package
