+++ sample/src/main/kotlin/io/getstream/chat/sample/feature/channel/ChannelFragment.kt
@@ -8,7 +8,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.getstream.sdk.chat.viewmodel.ChannelHeaderViewModel
 import com.getstream.sdk.chat.viewmodel.MessageInputViewModel
-import com.getstream.sdk.chat.viewmodel.MessageListViewModel
+import com.getstream.sdk.chat.viewmodel.messages.MessageListViewModel
 import com.getstream.sdk.chat.viewmodel.bindView
 import io.getstream.chat.sample.R
 import io.getstream.chat.sample.common.visible
